Update
Started writing html code for temperature resolution setting on the webpage.

+++ lab6_ek/DS1722.h
@@ -26,7 +26,7 @@
 void tempSetup(void);
 void setResolution(int res);
 int tempRead(void);
-int convertTemp(int bits);
+float convertTemp(int bits);
 
 #endif
